Refactor AniList provider into separate workspace
Move AniList provider from crates/providers/src/anilist/ to crates/providers/anilist/ as a separate workspace crate for better modularity.
