+++ crates/providers/anilist/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "anilist-provider"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+anyhow = { workspace = true }
+application-utils = { path = "../../utils/application" }
+async-trait = { workspace = true }
+chrono = { workspace = true }
+common-models = { path = "../../models/common" }
+common-utils = { path = "../../utils/common" }
+config = { path = "../../config" }
+convert_case = { workspace = true }
+dependent-models = { path = "../../models/dependent" }
+enum-models = { path = "../../models/enum" }
+itertools = { workspace = true }
+media-models = { path = "../../models/media" }
+nest_struct = { workspace = true }
+reqwest = { workspace = true }
+rust_decimal = { workspace = true }
+sea-orm = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+tracing = { workspace = true }
+traits = { path = "../../traits" }
+
+[package.metadata.cargo-machete]
+ignored = ["tracing"]
