+++ crates/providers/anilist/src/base.rs
