+++ crates/providers/anilist/src/lib.rs
