+++ crates/providers/anilist/src/models.rs
